wiped out the old `handlers.go`
